// fhf-datafangst-client/src/generated/openapi/models/tra.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { TraCatch } from './tra-catch';

/**
 * 
 * @export
 * @interface Tra
 */
export interface Tra {
    /**
     * 
     * @type {number}
     * @memberof Tra
     */
    'fiskeridirVesselId'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof Tra
     */
    'latitude'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof Tra
     */
    'longitude'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Tra
     */
    'reloadingTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Tra
     */
    'messageTimestamp': string;
    /**
     * 
     * @type {Array<TraCatch>}
     * @memberof Tra
     */
    'catches': Array<TraCatch>;
    /**
     * 
     * @type {number}
     * @memberof Tra
     */
    'reloadToFiskeridirVesselId'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof Tra
     */
    'reloadFromFiskeridirVesselId'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Tra
     */
    'reloadToCallSign'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Tra
     */
    'reloadFromCallSign'?: string | null;
}